+++ src/unfold/checks.py
@@ -4,12 +4,13 @@ from django.contrib.admin.checks import ModelAdminChecks
 from django.contrib.admin.options import BaseModelAdmin
 from django.contrib.auth.models import Permission
 from django.core import checks
+from django.core.checks import CheckMessage
 
 from unfold.dataclasses import UnfoldAction
 
 
 class UnfoldModelAdminChecks(ModelAdminChecks):
-    def check(self, admin_obj: BaseModelAdmin, **kwargs) -> list[checks.Error]:
+    def check(self, admin_obj: BaseModelAdmin, **kwargs) -> list[CheckMessage]:
         return [
             *super().check(admin_obj, **kwargs),
             *self._check_unfold_action_permission_methods(admin_obj),
@@ -42,7 +43,7 @@ class UnfoldModelAdminChecks(ModelAdminChecks):
                     ).exists():
                         errors.append(
                             checks.Error(
-                                f"@action decorator on {action.method.original_function_name}() in class {obj.__class__.__name__} specifies permission {permission} which does not exists.",
+                                f"@action decorator on {action.method.original_function_name}() in class {obj.__class__.__name__} specifies permission {permission} which does not exists.",  # type: ignore
                                 obj=obj.__class__,
                                 id="admin.E129",
                             )
@@ -55,7 +56,7 @@ class UnfoldModelAdminChecks(ModelAdminChecks):
                 if not hasattr(obj, method_name):
                     errors.append(
                         checks.Error(
-                            f"{obj.__class__.__name__} must define a {method_name}() method for the {action.method.original_function_name}() action.",
+                            f"{obj.__class__.__name__} must define a {method_name}() method for the {action.method.original_function_name}() action.",  # type: ignore
                             obj=obj.__class__,
                             id="admin.E129",
                         )