+++ tests/server/example/views.py
@@ -1,5 +1,8 @@
-from django.views.generic import TemplateView
+from django.urls import reverse_lazy
+from django.utils.translation import gettext_lazy as _
+from django.views.generic import FormView, TemplateView
 
+from example.forms import CrispyForm
 from unfold.views import UnfoldModelAdminViewMixin, UnfoldSiteViewMixin
 
 
@@ -19,3 +22,11 @@ class ModelExtraUrlView(UnfoldModelAdminViewMixin, TemplateView):
         "example.view_project",
         "example.add_project",
     )
+
+
+class CrispyFormView(UnfoldModelAdminViewMixin, FormView):
+    title = _("Crispy form")
+    success_url = reverse_lazy("admin:crispy_form")
+    permission_required = ("example.view_project",)
+    template_name = "crispy_form.html"
+    form_class = CrispyForm